Update test results and title in web pages
This commit updates the information of test results and title in two pages: index.html and ebw.html. In index.html, the date and time stamp of test results have been removed while in ebw.html, the title has been modified and the number of successful and failed tests has been updated with precise data.