Modify workflows to use force push
The 'updateFiles.yml' and 'lighthouse.yml' workflows have been updated to use 'git push -f' instead of 'git push'. This ensures that the commit history in the repository always reflects the most recent changes. The commit history is therefore easier to interpret, and potential merge conflicts because of out-of-date branches are avoided.