Remove assets/style.css
The assets folder has been cleaned up and unnecessary style sheets have been removed. This includes the total deletion of the "style.css" file which contained various styling configurations no longer required in the project. This step is part of the broader effort to refactor and streamline the project's assets.