Update file paths in updateFiles.yml workflow
The file paths in the .github/workflows/updateFiles.yml have been updated. The previous paths were directing towards the root, but have been changed to direct towards the 'pytest' directory instead. This change applies to both the 'index.html' file and the four other HTML files listed in the workflow.