Optimize date update in HTML report
The commit refactors the way the date gets updated in 'index.html'. Instead of manipulating the entire centred-text block in the html file, it now only adjusts the paragraph with the id 'date', which makes the operation considerably more efficient and less error-prone. This is achieved with 'sed' command in 'updateFiles.yml' workflow.